rename file: tex.tsx -> parse-result-map.tsx
